signup form, confirm message completed

+++ front/src/dashboard-section/DonarList.jsx
@@ -101,11 +101,16 @@ export default function Dashboard() {
   }, []);
 
   const handleDelete = async (id) => {
-    await axios.delete("http://127.0.0.1:8080/donar/" + id);
-    const newdonordata = data.filter((item) => {
-      return item._id !== id;
-    });
-    setData(newdonordata);
+    const confirmed = window.confirm(
+      "Are you sure you want to delete this item?"
+    );
+    if (confirmed) {
+      await axios.delete("http://127.0.0.1:8080/donar/" + id);
+      const newdonordata = data.filter((item) => {
+        return item._id !== id;
+      });
+      setData(newdonordata);
+    }
   };
 
   console.log(data);
@@ -364,7 +369,7 @@ export default function Dashboard() {
                               <td>{data?.email}</td>
                               <td>{data?.password}</td>
 
-                              <td class="flex space-x-4">
+                              <td className="flex space-x-4">
                                 <Link to={`/view/${data._id}`}>
                                   <svg
                                     xmlns="http://www.w3.org/2000/svg"
